feat: 회원 정보 필드 VO 구현
