+++ src/main/java/woowacourse/shoppingcart/domain/Email.java
@@ -1,0 +1,45 @@
+package woowacourse.shoppingcart.domain;
+
+import java.util.Objects;
+import java.util.regex.Pattern;
+import woowacourse.shoppingcart.exception.InvalidCustomerException;
+
+public class Email {
+
+    private static final Pattern EMAIL_PATTERN = Pattern.compile(
+            "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@[^-][A-Za-z0-9-]+(\\.[A-Za-z0-9-]+)*(\\.[A-Za-z]{2,})$");
+
+    private final String value;
+
+    public Email(String value) {
+        validateEmailFormat(value);
+        this.value = value;
+    }
+
+    private void validateEmailFormat(String email) {
+        if (!EMAIL_PATTERN.matcher(email).matches()) {
+            throw new InvalidCustomerException("잘못된 이메일 형식입니다.");
+        }
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Email email = (Email) o;
+        return Objects.equals(value, email.value);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(value);
+    }
+}
